added function to get integer types (like exp)
According to the JWT spec, exp is stored as an integer, not a string.
Other libraries create exp as an integer.  To read and validate that
the JWT is not expired, we have to read the exp.  Currently, this
library assumes all variables in the JWT are strings.  This provides an
integer getter.