Do not require typ to be present
Only validate typ if it's presend. RFC-7519/5.1 says it is optional.

Technically, I shouldn't even validate it if it exists, but there's no way
(yet) for the application to verify the typ. Since we only operate on JWT
types, anyway, we force validation to check it is typ="JWT" if it exists.

Fixes #37

Signed-off-by: Ben Collins <ben@cyphre.com>