Add signature verification function.
This commit adds a signature verification function,
jwt_check_signature(), which makes it possible that the
signature of a JWT can be verified not only during decoding,
in jwt_decode(), but also at a later stage, after decoding
has been otherwise successful. This may be necessary in
certain scenarios, where, in order to determine which
certificate to use for checking the signature, certain fields
of the decoded JWT need to be consulted (for example, the 'iss'
field) - in such a case the signature checking cannot be done
directly in jwt_decode().