jwt,examples: Improve example usage of JWT_ALG_INVAL check
Some functions can return this:

CORRECT:

	if (alg >= JWT_ALG_INVAL)
		/* handle problem */;

BAD:

	if (alg == JWT_ALG_INVAL)
		/* handle problem */;

Let's assume JWT_ALG_INVAL=8 when the program was compiled. Later, LibJWT
adds ES256K. Now JWT_ALG_ES256K=8 and JWT_ALG_INVAL=9. The code is looking
for a value of 8 to mean JWT_ALG_INVAL.

The second example above would not catch the error. It's sake for the
program to ignore JWT_ALG_ES256Kas well, since it doesn't know anything
about it until it is recompiled with the newer LibJWT.

Signed-off-by: Ben Collins <bcollins@maclara-llc.com>