gnutls: Piggyback off of OpenSSL's PEM creation
Since OpenSSL creates a PEM for each JWK it imports, we'll allow GnuTLS
to use that for now.

Secondly, GnuTLS assumed that for verifying, it required a public key.
The JWK spec allows private keys to support the "verify" key_op. GnuTLS
now checks if the key is private, and converts it to a pubkey in this
case.

This means OpenSSL is REQUIRED.

Signed-off-by: Ben Collins <bcollins@ubuntu.com>