gnutls: Fix signing with RSASSA-PSS
If JWKS imported an RSA key that did not have an algorithm set, OpenSSL
would store it's PEM file as a standard RSA, even if you wanted to use
it for RSASSA-PSS operations. GnuTLS wasn't working well with this, but
after a lot of trial and error, it came down to adding one single flags
to the sign_data() fn.

Signed-off-by: Ben Collins <bcollins@ubuntu.com>